# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Auth:
    properties:
      email:
        type: string
      hash:
        type: string
      login:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
    type: object
  response.Error:
    properties:
      code:
        type: integer
      details: {}
      message:
        type: string
    type: object
  response.Success:
    properties:
      data: {}
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:9090
info:
  contact:
    email: llchh@yahoo.com
    name: API Support
    url: gozakupki.com
  description: Documentation for Gozakupki api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger GoZakupki Api
  version: "1.0"
paths:
  /auth/check:
    get:
      consumes:
      - application/json
      description: Check if the authentication token is valid
      parameters:
      - default: Bearer
        description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Check authentication token
      tags:
      - Authentication
  /auth/confirm:
    post:
      consumes:
      - application/json
      description: Confirm a user's registration with the provided hash
      parameters:
      - description: Confirmation hash
        in: body
        name: hash
        required: true
        schema:
          $ref: '#/definitions/domain.Auth'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Confirm user registration
      tags:
      - Auth
  /auth/sign_in:
    post:
      consumes:
      - application/json
      description: Sign in with user credentials
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.Auth'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Sign in
      tags:
      - Authentication
  /auth/sign_up:
    post:
      consumes:
      - application/json
      description: Sign up with user credentials. Send hash to user email
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.Auth'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Sign up
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"